requirements.txt
aenum==3.1.15
asgiref==3.8.1
autopep8==2.3.2
cashfree-pg==4.3.10
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
Django==5.1.6
django-filter==24.3
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==36.1.0
flake8==7.1.2
gprof2dot==2024.6.6
httpie==3.2.4
idna==3.10
Markdown==3.7
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==1.10.21
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pyngrok==7.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
sentry-sdk==1.32.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.0.7
whitenoise
gunicorn